+++ LibCarla/source/carla/sensor/s11n/ImageSerializer.cpp
@@ -13,7 +13,12 @@ namespace sensor {
 namespace s11n {
 
   SharedPtr<SensorData> ImageSerializer::Deserialize(RawData data) {
-    return SharedPtr<data::Image>(new data::Image{std::move(data)});
+    auto image = SharedPtr<data::Image>(new data::Image{std::move(data)});
+    // Set alpha of each pixel in the buffer to max to make it 100% opaque
+    for (auto &pixel : *image) {
+      pixel.a = 255u;
+    }
+    return image;
   }
 
 } // namespace s11n