+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Latest
+  * Improved visual quality of the screen capture for the rgb sensor
+    - Enabled Temporal AA for screen captures with no post-processing to prevent jaggies during motion
+    - Reduced the target gamma of render target to 1.4 to minimize brightness differences with main camera 
   * Upgraded to Unreal Engine 4.22
   * Recorder fixes:
     - Fixed a possible crash if an actor is respawned before the episode is ready when a new map is loaded automatically.